Migrate custom types to types.py
