+++ src/quacc/recipes/espresso/phonons.py
@@ -23,18 +23,13 @@ from quacc.utils.dicts import recursive_dict_merge
 from quacc.wflow_tools.customizers import customize_funcs
 
 if TYPE_CHECKING:
-    from typing import Any, Callable, TypedDict
+    from typing import Any, Callable
 
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
-    from quacc.types import Filenames, SourceDirectory
+    from quacc.types import EspressoPhononDosSchema, Filenames, SourceDirectory
 
-    class PhononDosSchema(TypedDict):
-        relax_job: RunSchema
-        phonon_job: RunSchema
-        q2r_job: RunSchema
-        matdyn_job: RunSchema
 
 
 @job
@@ -204,7 +199,7 @@ def phonon_dos_flow(
     atoms: Atoms,
     job_params: dict[str, Any] | None = None,
     job_decorators: dict[str, Callable | None] | None = None,
-) -> PhononDosSchema:
+) -> EspressoPhononDosSchema:
     """
     Function to carry out a phonon DOS calculation. The phonon calculation is carried
     out on a coarse q-grid, the force constants are calculated and extrapolated to a
