+++ src/quacc/recipes/newtonnet/ts.py
@@ -28,19 +28,12 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
     from numpy.typing import NDArray
 
-    from quacc.recipes.newtonnet.core import FreqSchema
     from quacc.runners.ase import OptParams
-    from quacc.schemas._aliases.ase import OptSchema
-
-    class TSSchema(OptSchema):
-        freq_job: FreqSchema | None
-
-    class IRCSchema(OptSchema):
-        freq_job: FreqSchema | None
-
-    class QuasiIRCSchema(OptSchema):
-        irc_job: IRCSchema
-        freq_job: FreqSchema | None
+    from quacc.types import (
+        NewtonNetIRCSchema,
+        NewtonNetQuasiIRCSchema,
+        NewtonNetTSSchema,
+    )
 
 
 @job
@@ -55,7 +48,7 @@ def ts_job(
     freq_job_kwargs: dict[str, Any] | None = None,
     opt_params: OptParams | None = None,
     **calc_kwargs,
-) -> TSSchema:
+) -> NewtonNetTSSchema:
     """
     Perform a transition state (TS) job using the given atoms object.
 
@@ -134,7 +127,7 @@ def irc_job(
     freq_job_kwargs: dict[str, Any] | None = None,
     opt_params: OptParams | None = None,
     **calc_kwargs,
-) -> IRCSchema:
+) -> NewtonNetIRCSchema:
     """
     Perform an intrinsic reaction coordinate (IRC) job using the given atoms object.
 
@@ -213,7 +206,7 @@ def quasi_irc_job(
     irc_job_kwargs: dict[str, Any] | None = None,
     relax_job_kwargs: dict[str, Any] | None = None,
     freq_job_kwargs: dict[str, Any] | None = None,
-) -> QuasiIRCSchema:
+) -> NewtonNetQuasiIRCSchema:
     """
     Perform a quasi-IRC job using the given atoms object. The initial IRC job by default
     is run with `max_steps: 5`.
