--- src/quacc/schemas/_aliases/__init__.py
@@ -1,1 +1,0 @@
-"""Schemas for type hinting."""
