--- src/quacc/schemas/_aliases/atoms.py
@@ -1,19 +1,0 @@
-"""Aliases for type hinting `quacc.schemas.atoms`"""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-from quacc.schemas._aliases.emmet import MoleculeMetadata, StructureMetadata
-
-if TYPE_CHECKING:
-    from ase.atoms import Atoms
-    from pymatgen.core.structure import Molecule, Structure
-
-
-class AtomsSchema(StructureMetadata, MoleculeMetadata):
-    """Type hint associated with [quacc.schemas.atoms.atoms_to_metadata][]"""
-
-    atoms: Atoms
-    structure: Structure  # if atoms.pbc.any()
-    molecule: Molecule  # if not atoms.pbc.any()
